+++ web/embed.go
@@ -2,5 +2,5 @@ package web
 
 import "embed"
 
-//go:embed index.html memories.html status.html static
+//go:embed index.html status.html static
 var Static embed.FS